backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.36
alembic==1.14.1
psycopg2-binary==2.9.10
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.0
qrcode==7.4.2
pillow==11.1.0
pytest==7.4.4
httpx==0.26.0
python-jose[cryptography]==3.3.0
redis==5.0.1
fastapi-cache2==0.2.1
aioredis==2.0.1
email-validator==2.2.0 
APScheduler==3.11.0
